+++ Libplanet.Tests/Blockchain/BlockChainTest.Append.cs
@@ -126,7 +126,7 @@ namespace Libplanet.Tests.Blockchain
                 {
                     (Text)"foo", (Text)"bar", (Text)"baz", (Text)"qux", (Integer)1,
                 },
-                addresses.Select(renders[3].NextStates.GetState)
+                renders[3].NextStates.GetStates(addresses)
             );
 
             Address minerAddress = addresses[4];
@@ -164,11 +164,12 @@ namespace Libplanet.Tests.Blockchain
                 Assert.Equal(block2.Hash, s.BlockHash);
                 Assert.Equal(tx.Id, s.TxId);
                 Assert.Equal(tx.UpdatedAddresses, s.UpdatedAddresses);
+                Address[] updatedAddrs = tx.UpdatedAddresses.ToArray();
                 Assert.Equal(
-                    tx.UpdatedAddresses.ToImmutableDictionary(
-                        address => address,
-                        address => _blockChain.GetState(address)
-                    ),
+                    updatedAddrs.Zip(
+                        _blockChain.GetStates(updatedAddrs),
+                        (k, v) => new KeyValuePair<Address, IValue>(k, v)
+                    ).ToImmutableDictionary(),
                     s.UpdatedStates
                 );
                 Assert.Empty(s.FungibleAssetsDelta);