+++ Libplanet.Tests/Store/Trie/MerkleTrieTest.cs
@@ -62,10 +62,10 @@ namespace Libplanet.Tests.Store.Trie
                 FromString("1b16b1df538ba12dc3f97edbb85caa7050d46c148134290feba80f8236c83db9"),
                 trie.Hash
             );
-            Assert.False(trie.TryGet(new KeyBytes(0xbe, 0xef), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0x11, 0x22), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0xaa, 0xbb), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            Assert.Null(trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             trie = trie.Set(new KeyBytes(0xbe, 0xef), Null.Value);
             trie = commit ? trie.Commit() : trie;
@@ -73,11 +73,10 @@ namespace Libplanet.Tests.Store.Trie
                 FromString("16fc25f43edd0c2d2cb6e3cc3827576e57f4b9e04f8dc3a062c7fe59041f77bd"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out IValue got));
-            AssertBencodexEqual(Null.Value, got);
-            Assert.False(trie.TryGet(new KeyBytes(0x11, 0x22), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0xaa, 0xbb), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            AssertBencodexEqual(Null.Value, trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             trie = trie.Set(new KeyBytes(0xbe, 0xef), new Boolean(true));
             trie = commit ? trie.Commit() : trie;
@@ -85,11 +84,13 @@ namespace Libplanet.Tests.Store.Trie
                 FromString("4458796f4092b5ebfc1ffb3989e72edee228501e438080a12dea45591dc66d58"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.False(trie.TryGet(new KeyBytes(0x11, 0x22), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0xaa, 0xbb), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             trie = trie.Set(new KeyBytes(0x11, 0x22), List.Empty);
             trie = commit ? trie.Commit() : trie;
@@ -97,12 +98,13 @@ namespace Libplanet.Tests.Store.Trie
                 FromString("ab1359a2497453110a9c658dd3db45f282404fe68d8c8aca30856f395572284c"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(List.Empty, got);
-            Assert.False(trie.TryGet(new KeyBytes(0xaa, 0xbb), out _));
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(List.Empty, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             trie = trie.Set(new KeyBytes(0xaa, 0xbb), new Text("hello world"));
             trie = commit ? trie.Commit() : trie;
@@ -110,13 +112,16 @@ namespace Libplanet.Tests.Store.Trie
                 FromString("abb5759141f7af1c40f1b0993ba60073cf4227900617be9641373e5a097eaa3c"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(List.Empty, got);
-            Assert.True(trie.TryGet(new KeyBytes(0xaa, 0xbb), out got));
-            AssertBencodexEqual(new Text("hello world"), got);
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(List.Empty, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            AssertBencodexEqual(
+                new Text("hello world"),
+                trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]
+            );
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             var longText = new Text(string.Join("\n", Range(0, 1000).Select(i => $"long str {i}")));
             trie = trie.Set(new KeyBytes(0xaa, 0xbb), longText);
@@ -127,13 +132,13 @@ namespace Libplanet.Tests.Store.Trie
                     : "ad9fb53a8f643bd308d7afea57a5d1796d6031b1df95bdd415fa69b44177d155"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(List.Empty, got);
-            Assert.True(trie.TryGet(new KeyBytes(0xaa, 0xbb), out got));
-            AssertBencodexEqual(longText, got);
-            Assert.False(trie.TryGet(new KeyBytes(0x12, 0x34), out _));
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(List.Empty, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            AssertBencodexEqual(longText, trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            Assert.Null(trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             trie = trie.Set(new KeyBytes(0x12, 0x34), Dictionary.Empty);
             trie = commit ? trie.Commit() : trie;
@@ -143,14 +148,13 @@ namespace Libplanet.Tests.Store.Trie
                     : "77d13e9d97033400ad31fcb0441819285b9165f6ea6ae599d85e7d7e24428feb"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(List.Empty, got);
-            Assert.True(trie.TryGet(new KeyBytes(0xaa, 0xbb), out got));
-            AssertBencodexEqual(longText, got);
-            Assert.True(trie.TryGet(new KeyBytes(0x12, 0x34), out got));
-            AssertBencodexEqual(Dictionary.Empty, got);
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(List.Empty, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            AssertBencodexEqual(longText, trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            AssertBencodexEqual(Dictionary.Empty, trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             List complexList = List.Empty
                 .Add("Hello world")
@@ -169,14 +173,13 @@ namespace Libplanet.Tests.Store.Trie
                     : "586ba0ba5dfe07433b01fbf7611f95832bde07b8dc5669540ef8866f465bbb85"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(complexList, got);
-            Assert.True(trie.TryGet(new KeyBytes(0xaa, 0xbb), out got));
-            AssertBencodexEqual(longText, got);
-            Assert.True(trie.TryGet(new KeyBytes(0x12, 0x34), out got));
-            AssertBencodexEqual(Dictionary.Empty, got);
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(complexList, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            AssertBencodexEqual(longText, trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            AssertBencodexEqual(Dictionary.Empty, trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
 
             Dictionary complexDict = Dictionary.Empty
                 .Add("foo", 123)
@@ -200,14 +203,13 @@ namespace Libplanet.Tests.Store.Trie
                     : "4783d18dfc8a2d4d98f722a935e45bd7fc1d0197fb4d33e62f734bfde968af39"),
                 trie.Hash
             );
-            Assert.True(trie.TryGet(new KeyBytes(0xbe, 0xef), out got));
-            AssertBencodexEqual(new Boolean(true), got);
-            Assert.True(trie.TryGet(new KeyBytes(0x11, 0x22), out got));
-            AssertBencodexEqual(complexList, got);
-            Assert.True(trie.TryGet(new KeyBytes(0xaa, 0xbb), out got));
-            AssertBencodexEqual(longText, got);
-            Assert.True(trie.TryGet(new KeyBytes(0x12, 0x34), out got));
-            AssertBencodexEqual(complexDict, got);
+            AssertBencodexEqual(
+                new Boolean(true),
+                trie.Get(new[] { new KeyBytes(0xbe, 0xef) })[0]
+            );
+            AssertBencodexEqual(complexList, trie.Get(new[] { new KeyBytes(0x11, 0x22) })[0]);
+            AssertBencodexEqual(longText, trie.Get(new[] { new KeyBytes(0xaa, 0xbb) })[0]);
+            AssertBencodexEqual(complexDict, trie.Get(new[] { new KeyBytes(0x12, 0x34) })[0]);
         }
     }
 }