+++ Libplanet/Blockchain/BlockChain.cs
@@ -33,7 +33,7 @@ namespace Libplanet.Blockchain
     /// block when it's instantiated.</remarks>
     /// <typeparam name="T">An <see cref="IAction"/> type.  It should match
     /// to <see cref="Block{T}"/>'s type parameter.</typeparam>
-    public partial class BlockChain<T>
+    public partial class BlockChain<T> : IBlockChainStates<T>
         where T : IAction, new()
     {
         // FIXME: The _rwlock field should be private.
@@ -164,9 +164,8 @@ namespace Libplanet.Blockchain
                 .ForContext("CanonicalChainId", Id);
             ActionEvaluator = new ActionEvaluator<T>(
                 Policy.BlockAction,
-                GetState,
-                GetBalance,
-                hash => StateStore.GetStateRoot(_blocks[hash].StateRootHash)
+                blockChainStates: this,
+                trieGetter: hash => StateStore.GetStateRoot(_blocks[hash].StateRootHash)
             );
 
             if (Count == 0)
@@ -470,14 +469,61 @@ namespace Libplanet.Blockchain
             BlockHash? offset = null,
             StateCompleter<T> stateCompleter = null
         ) =>
-            GetRawState(
-                ToStateKey(address),
-                offset,
-                StateCompleters<T>.ToRawStateCompleter(
-                    stateCompleter ?? StateCompleters<T>.Reject,
-                    address
-                )
+            GetStates(
+                new[] { address },
+                offset ?? Tip.Hash,
+                stateCompleter ?? StateCompleters<T>.Reject
+            )[0];
+
+        /// <summary>
+        /// Gets multiple states associated to the specified <paramref name="addresses"/>.
+        /// </summary>
+        /// <param name="addresses">Addresses of states to query.</param>
+        /// <param name="offset">The <see cref="HashDigest{T}"/> of the block to start finding
+        /// the states.  <see cref="Tip"/> by default.</param>
+        /// <param name="stateCompleter">When the <see cref="BlockChain{T}"/> instance does not
+        /// contain states of the block, this delegate is called and its return values are used
+        /// instead.
+        /// <para><see cref="StateCompleters{T}.Recalculate"/> makes the incomplete states
+        /// recalculated and filled on the fly.</para>
+        /// <para><see cref="StateCompleters{T}.Reject"/> (which is default) makes the incomplete
+        /// states (if needed) to cause <see cref="IncompleteBlockStatesException"/> instead.</para>
+        /// </param>
+        /// <returns>The states associated to the specified <paramref name="addresses"/>.
+        /// Associated values are ordered in the same way to the corresponding
+        /// <paramref name="addresses"/>.  Absent states are represented as <see langword="null"/>.
+        /// </returns>
+        public IReadOnlyList<IValue> GetStates(
+            IReadOnlyList<Address> addresses,
+            BlockHash? offset = null,
+            StateCompleter<T> stateCompleter = null
+        ) =>
+            GetStates(
+                addresses,
+                offset ?? Tip.Hash,
+                stateCompleter ?? StateCompleters<T>.Reject
             );
+
+        public IReadOnlyList<IValue> GetStates(
+            IReadOnlyList<Address> addresses,
+            BlockHash offset,
+            StateCompleter<T> stateCompleter
+        )
+        {
+            if (Count < 1)
+            {
+                return new IValue[addresses.Count];
+            }
+
+            HashDigest<SHA256>? stateRootHash = Store.GetStateRootHash(offset);
+            if (stateRootHash is { } h && StateStore.ContainsStateRoot(h))
+            {
+                string[] rawKeys = addresses.Select(ToStateKey).ToArray();
+                return StateStore.GetStates(stateRootHash, rawKeys);
+            }
+
+            return stateCompleter(this, offset, addresses);
+        }
 
         /// <summary>
         /// Queries <paramref name="address"/>'s balance of the <paramref name="currency"/> in the
@@ -504,22 +550,39 @@ namespace Libplanet.Blockchain
             Currency currency,
             BlockHash? offset = null,
             FungibleAssetStateCompleter<T> stateCompleter = null
+        ) =>
+            GetBalance(
+                address,
+                currency,
+                offset ?? Tip.Hash,
+                stateCompleter ?? FungibleAssetStateCompleters<T>.Reject
+            );
+
+        /// <inheritdoc cref="IBlockChainStates{T}.GetBalance"/>
+        public FungibleAssetValue GetBalance(
+            Address address,
+            Currency currency,
+            BlockHash offset,
+            FungibleAssetStateCompleter<T> stateCompleter
         )
         {
-            stateCompleter ??= FungibleAssetStateCompleters<T>.Reject;
-            IValue v = GetRawState(
-                ToFungibleAssetKey(address, currency),
-                offset,
-                FungibleAssetStateCompleters<T>.ToRawStateCompleter(
-                    stateCompleter,
-                    address,
-                    currency
-                )
-            );
-            return FungibleAssetValue.FromRawValue(
-                currency,
-                v is Bencodex.Types.Integer i ? i.Value : 0
-            );
+            if (Count < 1)
+            {
+                return currency * 0;
+            }
+
+            HashDigest<SHA256>? stateRootHash = Store.GetStateRootHash(offset);
+            if (stateRootHash is { } h && StateStore.ContainsStateRoot(h))
+            {
+                string rawKey = ToFungibleAssetKey(address, currency);
+                IReadOnlyList<IValue> values =
+                    StateStore.GetStates(stateRootHash, new[] { rawKey });
+                return values.Count > 0 && values[0] is Bencodex.Types.Integer i
+                    ? FungibleAssetValue.FromRawValue(currency, i)
+                    : currency * 0;
+            }
+
+            return stateCompleter(this, offset, address, currency);
         }
 
         /// <summary>
@@ -1357,25 +1420,6 @@ namespace Libplanet.Blockchain
             }
 
             return null;
-        }
-
-        private IValue GetRawState(
-            string key,
-            BlockHash? offset,
-            Func<BlockChain<T>, BlockHash, IValue> rawStateCompleter
-        )
-        {
-            if (offset is null && Count == 0)
-            {
-                return null;
-            }
-
-            BlockHash offsetHash = offset ?? Tip.Hash;
-
-            HashDigest<SHA256>? stateRootHash = Store.GetStateRootHash(offsetHash);
-            return stateRootHash is { } h && StateStore.ContainsStateRoot(h)
-                ? StateStore.GetState(key, stateRootHash)
-                : rawStateCompleter(this, offsetHash);
         }
     }
 }